+++ rasa_sdk/endpoint.py
@@ -1,21 +1,38 @@
 import argparse
 import logging
+import types
+from typing import List, Text, Union, Optional, Any
+from ssl import SSLContext
 
-from flask import Flask, jsonify, request
-from flask_cors import CORS, cross_origin
-from gevent.pywsgi import WSGIServer
+from sanic import Sanic, response
+from sanic.response import HTTPResponse
+from sanic.request import Request
+from sanic_cors import CORS
 
-import rasa_sdk
 from rasa_sdk import utils
 from rasa_sdk.cli.arguments import add_endpoint_arguments
 from rasa_sdk.constants import DEFAULT_SERVER_PORT
 from rasa_sdk.executor import ActionExecutor
-from rasa_sdk.interfaces import ActionExecutionRejection
+from rasa_sdk.interfaces import ActionExecutionRejection, ActionNotFoundException
 
 logger = logging.getLogger(__name__)
 
 
-def create_ssl_context(ssl_certificate, ssl_keyfile, ssl_password):
+def configure_cors(
+    app: Sanic, cors_origins: Union[Text, List[Text], None] = ""
+) -> None:
+    """Configure CORS origins for the given app."""
+
+    CORS(
+        app, resources={r"/*": {"origins": cors_origins or ""}}, automatic_options=True
+    )
+
+
+def create_ssl_context(
+    ssl_certificate: Optional[Text],
+    ssl_keyfile: Optional[Text],
+    ssl_password: Optional[Text] = None,
+) -> Optional[SSLContext]:
     """Create a SSL context if a certificate is passed."""
 
     if ssl_certificate:
@@ -39,120 +56,67 @@ def create_argument_parser():
     return parser
 
 
-def endpoint_app(cors_origins=None, action_package_name=None):
-    app = Flask(__name__)
+def create_app(
+    action_package_name: Union[Text, types.ModuleType],
+    cors_origins: Union[Text, List[Text], None] = "*",
+) -> Sanic:
+    app = Sanic(__name__, configure_logging=False)
 
-    if not cors_origins:
-        cors_origins = []
+    configure_cors(app, cors_origins)
 
     executor = ActionExecutor()
     executor.register_package(action_package_name)
 
-    CORS(app, resources={r"/*": {"origins": cors_origins}})
-
-    @app.route("/health", methods=["GET", "OPTIONS"])
-    @cross_origin(origins=cors_origins)
-    def health():
+    @app.get("/health")
+    async def health(_) -> HTTPResponse:
         """Ping endpoint to check if the server is running and well."""
-        return jsonify({"status": "ok"})
+        body = {"status": "ok"}
+        return response.json(body, status=200)
 
-    @app.route("/webhook", methods=["POST", "OPTIONS"])
-    @cross_origin()
-    def webhook():
+    @app.post("/webhook")
+    async def webhook(request: Request) -> HTTPResponse:
         """Webhook to retrieve action calls."""
         action_call = request.json
-        check_version_compatibility(action_call.get("version"))
+        if action_call is None:
+            body = {"error": "Invalid body request"}
+            return response.json(body, status=400)
+
+        utils.check_version_compatibility(action_call.get("version"))
         try:
-            response = executor.run(action_call)
+            result = await executor.run(action_call)
         except ActionExecutionRejection as e:
-            logger.error(str(e))
-            result = {"error": str(e), "action_name": e.action_name}
-            response = jsonify(result)
-            response.status_code = 400
-            return response
+            logger.error(e)
+            body = {"error": e.message, "action_name": e.action_name}
+            return response.json(body, status=400)
+        except ActionNotFoundException as e:
+            logger.error(e)
+            body = {"error": e.message, "action_name": e.action_name}
+            return response.json(body, status=404)
 
-        return jsonify(response)
+        return response.json(result, status=200)
 
-    @app.route("/actions", methods=["GET", "OPTIONS"])
-    @cross_origin(origins=cors_origins)
-    def actions():
+    @app.get("/actions")
+    async def actions(_) -> HTTPResponse:
         """List all registered actions."""
-        return jsonify([{"name": k} for k in executor.actions.keys()])
+        body = [{"name": k} for k in executor.actions.keys()]
+        return response.json(body, status=200)
 
     return app
 
 
-def check_version_compatibility(rasa_version):
-    """Check if the version of rasa and rasa_sdk are compatible.
-
-    The version check relies on the version string being formatted as
-    'x.y.z' and compares whether the numbers x and y are the same for both
-    rasa and rasa_sdk.
-    Args:
-        rasa_version - A string containing the version of rasa that
-        is making the call to the action server.
-    Raises:
-        Warning - The version of rasa version unknown or not compatible with
-        this version of rasa_sdk.
-    """
-    # Check for versions of Rasa that are too old to report their version number
-    if rasa_version is None:
-        logger.warning(
-            "You are using an old version of rasa which might "
-            "not be compatible with this version of rasa_sdk "
-            "({}).\n"
-            "To ensure compatibility use the same version "
-            "for both, modulo the last number, i.e. using version "
-            "A.B.x the numbers A and B should be identical for "
-            "both rasa and rasa_sdk."
-            "".format(rasa_sdk.__version__)
-        )
-        return
-
-    rasa = rasa_version.split(".")[:-1]
-    sdk = rasa_sdk.__version__.split(".")[:-1]
-
-    if rasa != sdk:
-        logger.warning(
-            "Your versions of rasa and "
-            "rasa_sdk might not be compatible. You "
-            "are currently running rasa version {} "
-            "and rasa_sdk version {}.\n"
-            "To ensure compatibility use the same "
-            "version for both, modulo the last number, "
-            "i.e. using version A.B.x the numbers A and "
-            "B should be identical for "
-            "both rasa and rasa_sdk."
-            "".format(rasa_version, rasa_sdk.__version__)
-        )
-
-
 def run(
-    action_package_name,
-    port=DEFAULT_SERVER_PORT,
-    cors_origins="*",
-    ssl_certificate=None,
-    ssl_keyfile=None,
-    ssl_password=None,
+    action_package_name: Union[Text, types.ModuleType],
+    port: Union[Text, int] = DEFAULT_SERVER_PORT,
+    cors_origins: Union[Text, List[Text], None] = "*",
+    ssl_certificate: Optional[Text] = None,
+    ssl_keyfile: Optional[Text] = None,
+    ssl_password: Optional[Text] = None,
 ):
     logger.info("Starting action endpoint server...")
-    edp_app = endpoint_app(
-        cors_origins=cors_origins, action_package_name=action_package_name
-    )
+    app = create_app(action_package_name, cors_origins=cors_origins)
     ssl_context = create_ssl_context(ssl_certificate, ssl_keyfile, ssl_password)
-    protocol = "https" if ssl_context else "http"
-    if ssl_context:
-        http_server = WSGIServer(("0.0.0.0", port), edp_app, ssl_context=ssl_context)
-    else:
-        http_server = WSGIServer(("0.0.0.0", port), edp_app)
-    http_server.start()
-    logger.info(
-        "Action endpoint is up and running on {} {}".format(
-            protocol, http_server.address
-        )
-    )
 
-    http_server.serve_forever()
+    app.run("0.0.0.0", port, ssl=ssl_context, workers=utils.number_of_sanic_workers())
 
 
 if __name__ == "__main__":