+++ rasa_sdk/utils.py
@@ -1,6 +1,14 @@
 import inspect
 import logging
-from typing import Any, List, Text
+import warnings
+import os
+
+from typing import Any, List, Text, Optional
+
+import rasa_sdk
+from rasa_sdk.constants import DEFAULT_SANIC_WORKERS, ENV_SANIC_WORKERS
+
+logger = logging.getLogger(__name__)
 
 
 class Element(dict):
@@ -76,3 +84,86 @@ def arguments_of(func) -> List[Text]:
     """Return the parameters of the function `func` as a list of their names."""
 
     return inspect.signature(func).parameters.keys()
+
+
+def number_of_sanic_workers() -> int:
+    """Get the number of Sanic workers to use in `app.run()`.
+    If the environment variable `constants.ENV_SANIC_WORKERS` is set and is not equal to 1.
+    """
+
+    def _log_and_get_default_number_of_workers():
+        logger.debug(
+            f"Using the default number of Sanic workers ({DEFAULT_SANIC_WORKERS})."
+        )
+        return DEFAULT_SANIC_WORKERS
+
+    try:
+        env_value = int(os.environ.get(ENV_SANIC_WORKERS, DEFAULT_SANIC_WORKERS))
+    except ValueError:
+        logger.error(
+            f"Cannot convert environment variable `{ENV_SANIC_WORKERS}` "
+            f"to int ('{os.environ[ENV_SANIC_WORKERS]}')."
+        )
+        return _log_and_get_default_number_of_workers()
+
+    if env_value == DEFAULT_SANIC_WORKERS:
+        return _log_and_get_default_number_of_workers()
+
+    if env_value < 1:
+        warnings.warn(
+            f"Cannot set number of Sanic workers to the desired value "
+            f"({env_value}). The number of workers must be at least 1."
+        )
+        return _log_and_get_default_number_of_workers()
+
+    logger.debug(f"Using {env_value} Sanic workers.")
+    return env_value
+
+
+def check_version_compatibility(rasa_version: Optional[Text]) -> None:
+    """Check if the version of rasa and rasa_sdk are compatible.
+
+    The version check relies on the version string being formatted as
+    'x.y.z' and compares whether the numbers x and y are the same for both
+    rasa and rasa_sdk.
+    Args:
+        rasa_version - A string containing the version of rasa that
+        is making the call to the action server.
+    Raises:
+        Warning - The version of rasa version unknown or not compatible with
+        this version of rasa_sdk.
+    """
+    # Check for versions of Rasa that are too old to report their version number
+    if rasa_version is None:
+        warnings.warn(
+            "You are using an old version of rasa which might "
+            "not be compatible with this version of rasa_sdk "
+            "({}).\n"
+            "To ensure compatibility use the same version "
+            "for both, modulo the last number, i.e. using version "
+            "A.B.x the numbers A and B should be identical for "
+            "both rasa and rasa_sdk."
+            "".format(rasa_sdk.__version__)
+        )
+        return
+
+    rasa = rasa_version.split(".")[:-1]
+    sdk = rasa_sdk.__version__.split(".")[:-1]
+
+    if rasa != sdk:
+        warnings.warn(
+            "Your versions of rasa and "
+            "rasa_sdk might not be compatible. You "
+            "are currently running rasa version {} "
+            "and rasa_sdk version {}.\n"
+            "To ensure compatibility use the same "
+            "version for both, modulo the last number, "
+            "i.e. using version A.B.x the numbers A and "
+            "B should be identical for "
+            "both rasa and rasa_sdk."
+            "".format(rasa_version, rasa_sdk.__version__)
+        )
+
+
+def is_coroutine_action(action) -> bool:
+    return inspect.iscoroutinefunction(action)