+++ tests/test_endpoint.py
@@ -1,6 +1,55 @@
 import pytest
+import json
 
 import rasa_sdk.endpoint as ep
+from rasa_sdk.events import SlotSet
+
+app = ep.create_app("actions.act")
+
+
+def test_server_health_returns_200():
+    request, response = app.test_client.get("/health")
+    assert response.status == 200
+    assert response.json == {"status": "ok"}
+
+
+def test_server_list_actions_returns_200():
+    request, response = app.test_client.get("/actions")
+    assert response.status == 200
+    assert len(response.json) == 2
+
+
+def test_server_webhook_unknown_action_returns_404():
+    data = {
+        "next_action": "test_action_1",
+        "tracker": {"sender_id": "1", "conversation_id": "default"},
+    }
+    request, response = app.test_client.post("/webhook", data=json.dumps(data))
+    assert response.status == 404
+
+
+def test_server_webhook_custom_action_returns_200():
+    data = {
+        "next_action": "custom_action",
+        "tracker": {"sender_id": "1", "conversation_id": "default"},
+    }
+    request, response = app.test_client.post("/webhook", data=json.dumps(data))
+    events = response.json.get("events")
+
+    assert events == [SlotSet("test", "test")]
+    assert response.status == 200
+
+
+def test_server_webhook_custom_async_action_returns_200():
+    data = {
+        "next_action": "custom_async_action",
+        "tracker": {"sender_id": "1", "conversation_id": "default"},
+    }
+    request, response = app.test_client.post("/webhook", data=json.dumps(data))
+    events = response.json.get("events")
+
+    assert events == [SlotSet("test", "foo"), SlotSet("test2", "boo")]
+    assert response.status == 200
 
 
 def test_arg_parser_actions_params_folder_style():